+++ client/my-sites/plugins/plugin-management-v2/plugin-list-v2/style.scss
@@ -1,0 +1,26 @@
+
+body.is-section-plugins .plugin-management-wrapper {
+
+    .dataviews-wrapper {
+        .dataviews-view-table {
+            border: 1px solid var(--color-neutral-5);
+            border-radius: 2px;
+
+            .dataviews-view-table__row {
+                img, .dataviews-view-table__row svg {
+                    max-width: 35px;
+                    margin-right: 15px;
+                }
+
+                .dataviews-view-table-selection-checkbox{
+                    padding-left: 0;
+                }
+            }
+
+            .dataviews-view-table__checkbox-column {
+                padding-left: 18px;
+            }
+        }
+    }
+
+}
