+++ client/my-sites/plugins/style.scss
@@ -523,32 +523,6 @@ body.is-section-plugins .plugin-management-wrapper {
 		padding-left: 2rem;
 		padding-right: 2rem;
 	}
-
-	.dataviews-wrapper {
-		.dataviews__view-actions {
-			display: none;
-		}
-		.dataviews-view-table {
-			border: 1px solid var(--color-neutral-5);
-			border-radius: 2px;
-	
-			.dataviews-view-table__row {
-				img, .dataviews-view-table__row svg {
-					max-width: 35px;
-					margin-right: 15px;
-				}
-	
-				.dataviews-view-table-selection-checkbox{
-					padding-left: 0;
-				}
-			}
-	
-			.dataviews-view-table__checkbox-column {
-				padding-left: 18px;
-			}
-		}
-	}
-
 }
 
 .plugins__page-title {
